changed int_to_float64 to binary.py
